misc: mediatek: accdet: mt6763: We're not using Python here, this is C
As the driver was being compiled with Werror, the compilation died here.

This one was especially bad as it looks like the two lines should've
always been executed together under the if-clause, but only the first
one ever was while the second one was being executed always.

Every other instance where eint_type was used like this, or
clear_accdet_interrupt() was being called in had a similar if-clause
already in place.

Oh and also remove the useless double-parenthesis, these are not needed.

Fixes the following error during compilation:

  drivers/misc/mediatek/accdet/mt6763/accdet.c: In function 'accdet_irq_handler':
  drivers/misc/mediatek/accdet/mt6763/accdet.c:1416:2: error: this 'if' clause does not guard... [-Werror=misleading-indentation]
   1416 |  if ((pmic_pwrap_read(ACCDET_CON12) & ACCDET_IRQ_B0))
        |  ^~
  drivers/misc/mediatek/accdet/mt6763/accdet.c:1418:3: note: ...this statement, but the latter is misleadingly indented as if it were guarded by the 'if'
   1418 |   clear_accdet_interrupt();
        |   ^~~~~~~~~~~~~~~~~~~~~~

Signed-off-by: Jami Kettunen <jami.kettunen@protonmail.com>